requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
elasticsearch==9.1.0
opensearch-py==3.0.0
python-dotenv==1.1.1
nltk==3.9.1
pillow==11.3.0
pytesseract==0.3.13
